+++ virtualizarr/accessor.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 from typing import (
+    TYPE_CHECKING,
     Callable,
     Literal,
     overload,
@@ -11,6 +12,9 @@ from virtualizarr.manifests import ManifestArray
 from virtualizarr.types.kerchunk import KerchunkStoreRefs
 from virtualizarr.writers.kerchunk import dataset_to_kerchunk_refs
 from virtualizarr.writers.zarr import dataset_to_zarr
+
+if TYPE_CHECKING:
+    from icechunk import IcechunkStore  # type: ignore[import-not-found]
 
 
 @register_dataset_accessor("virtualize")
@@ -38,6 +42,20 @@ class VirtualiZarrDatasetAccessor:
         storepath : str
         """
         dataset_to_zarr(self.ds, storepath)
+
+    def to_icechunk(self, store: "IcechunkStore") -> None:
+        """
+        Write an xarray dataset to an Icechunk store.
+
+        Any variables backed by ManifestArray objects will be be written as virtual references, any other variables will be loaded into memory before their binary chunk data is written into the store.
+
+        Parameters
+        ----------
+        store: IcechunkStore
+        """
+        from virtualizarr.writers.icechunk import dataset_to_icechunk
+
+        dataset_to_icechunk(self.ds, store)
 
     @overload
     def to_kerchunk(